adding two examples notebooks and reorganizing examples folder
this removes the working notebooks and adds two use cases notebooks. The NLCD_example notebook is a step-by-step notebook with an end-to-end process of the NLCD interpolation using the functions built. The notebook NLCD_weights_from_regreession_examplained_example is a notebook that explains in more detail what is happening in every function of this module, specially in terms of the GeoProcessing, transformations, etc.